noInterface, kafkaRefactor, comment and envVariable
