+++ src/main/resources/application.properties
@@ -12,4 +12,4 @@ camunda.server.engine=${CAMUNDA_ENGINE_REST:http://localhost:8080/engine-rest/}
 camunda.bpm.client.disable-backoff-strategy=true
 spring.main.allow-bean-definition-overriding=true
 
-spring.kafka.bootstrap-servers=localhost:9092
+spring.kafka.bootstrap-servers=${KAFKA_BOOTSTRAP_SERVERS:localhost:9092}
